+++ source/common-js/index.js
@@ -1,2 +1,3 @@
+import "../common-js/md5.js";
 import "../common-js/utilities.js";
 import "../components/carousel/animations.js";