+++ source/components/gravatar/index.js
@@ -1,3 +1,3 @@
-import "../md5/index.js";
+import "../../common-js/md5.js";
 import "./gravatar.less";
 import "./gravatar.js";