+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 + [x] Added event `onOptions` to `select` component.
 + [x] To remote data in `select` component added `selected` prop.
++ [x] `md5` moved to `common-js`